+++ SCPSLAudioApi/SCPSLAudioApi.csproj
@@ -227,7 +227,8 @@
     </ItemGroup>
 
     <ItemGroup>
-        <Compile Include="AudioCore\AudioPlayerBase.cs"/>
+        <Compile Include="Api\AudioPlayerBase.cs"/>
+        <Compile Include="AudioCore\AudioPlayerComponent.cs"/>
         <Compile Include="Events\Handlers\Track.cs"/>
         <Compile Include="Startup.cs"/>
     </ItemGroup>
@@ -240,10 +241,6 @@
     <ItemGroup>
         <Content Include="FodyWeavers.xml"/>
         <Content Include="FodyWeavers.xsd"/>
-    </ItemGroup>
-
-    <ItemGroup>
-        <Folder Include="Api"/>
     </ItemGroup>
 
     <Import Project="..\packages\Fody.6.6.0\build\Fody.targets" Condition="Exists('..\packages\Fody.6.6.0\build\Fody.targets')"/>
